+++ src/flavors/expo-pkg/expo-pkg.ts
@@ -1,15 +1,15 @@
 import type { FlavorFunction } from '../../typesAndConsts';
 import editJsonFile from 'edit-json-file';
+import { Core } from '../../core';
+import ora from 'ora';
 
+export const flavorExpoPkg: FlavorFunction = async (core) => {
 
-// Add Heroku
-
-
-export const flavorHapi: FlavorFunction = async (core) => {
+  core.verifications.projectNameMustBeNpmValid();
 
   await core.verifications.projectPathMustBeValid();
 
-  console.log(`Generating the hapi project "${core.consts.projectName}" at "${core.consts.projectPath}"...`);
+  ora().info(`Generating the Expo Library package "${core.consts.projectName}" at "${core.consts.projectPath}"`);
 
   await core.actions.setProjectDirectory();
 
@@ -29,19 +29,29 @@ export const flavorHapi: FlavorFunction = async (core) => {
   await core.actions.addPackages({
     devDeps: [
       'typescript@latest',
-
       'eslint@latest',
       'eslint-config-gev@latest',
       'eslint-plugin-react@latest',
       '@typescript-eslint/eslint-plugin@latest',
       '@typescript-eslint/parser@latest',
-
-      '@types/hapi__hapi',
-    ],
-    deps: [
-      '@hapi/hapi@latest',
     ],
   });
 
-  console.log(`hapi project "${core.consts.projectName}" created at "${core.consts.projectPath}"!`);
+  ora().info('Adding sandbox directory, via gev expo');
+
+
+  const sandboxCore = new Core({
+    cwd: core.consts.projectPath,
+    receivedProjectName: 'sandbox',
+    flavor: 'expo',
+    installPackages: core.consts.installPackages,
+    cleanOnError: core.consts.cleanOnError,
+  });
+  await sandboxCore.run();
+
+  ora().succeed(`expo-pkg package "${core.consts.projectName}" created at "${core.consts.projectPath}"!`);
+
+  // fse.symlink() // symlink to src inside src.
+  // in expo, should add watch to the ../src.
 };
+